First upload
Starts from Tester_Select, which selects the application and starts it.
The visual window is created in Frame, which has the functions that keep
track of periodic updating (by timer) and redrawing of the frame. It has
(currently) all the handlers (one handler for each type of
moving/stationary object, which contains all objects of that type).  It
contains objects Graphics, which (will) be responsible for how the frame
is drawn, and Collisions, which checks collisions for all objects in
handlers.