Collisons and graphics
HandlerTracker now contains all handlers (which have all physical
objects). Class Graphics get the HandlerTracker, and tells all handlers
to draw their objects. Events works in a similar fashion, for
keyboard/mouse key interrupt events. Collisions get all physical objects
from the HandlerTracker and do a collision check. It is quite buggy, and
the middle of the images collide with the boundaries of other images, so
they start vibrating (this needs to be fixed by colliding boundary with
boundary). Also, adding a new Class of physical collidable objects now
requires instanceof checks outside of the HandlerTracker, which is
messy. This needs to be redone, so that all physical objects extend some
CollisionObject class, that contains all physics info of the physical
object (e.g. to do collisions). Also this highlights a problem with
having one handler per object type, because this requires instanceof
asserting both of the handler, and the object it contains. It might just
be better to not have a HandlerTracker, and just keep all physical
object in 1 handler (instead of a tracker of handlers).